tests/wpt/tests/css/css-multicol/multicol-fill-balance-019.html

<!DOCTYPE html>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://www.w3.org/TR/css-multicol-1/#cf">
<div id="mc" style="columns:2; column-gap:0; width:100px; orphans:1; widows:1; line-height:20px;">
  <span>
    <div style="height:200px;"></div>
  </span>
</div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(()=> {
    assert_equals(mc.offsetHeight, 100);
  }, "Breakable block inside inline");
</script>
